Added the feature of sunflower producing sun at regular intervals

+++ src/PvZ/Character.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public abstract class Character extends GraphicObject implements Serializable {
     protected int HP;
     protected final int attack_power;
-    public Character(int HP,int attack_power,String img,int x,int y){
+    public Character(int HP,int attack_power,String img,double x,double y){
         super(img,x,y);
         this.HP=HP;
         this.attack_power=attack_power;
