+++ src/PvZ/Tripeater.java
@@ -3,7 +3,7 @@ package PvZ;
 import java.io.Serializable;
 
 public class Tripeater extends Plant implements Serializable {
-    public Tripeater(int x,int y){
+    public Tripeater(double x,double y){
         super(300,100,100,10,x,y,"pea_shooter.gif");
     }
 }
